Use create_host_file in run -v tests to ensure file availability
